% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abort_packages_not_installed}
\alias{abort_packages_not_installed}
\title{Throw error if required packages are not installed.}
\usage{
abort_packages_not_installed(...)
}
\arguments{
\item{...}{names of packages to check}
}
\description{
Reports which packages need to be installed and the parent function name.
https://stackoverflow.com/questions/15595478/how-to-get-the-name-of-the-calling-function-inside-the-called-routine
}
\details{
This is only intended to be used inside a function. It will error otherwise.
}
\examples{
\dontrun{
abort_packages_not_installed("base")
abort_packages_not_installed("not-a-package-name", "caret", "dplyr", "non_package")
}
}
\keyword{internal}
