% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prelim_stats.R
\name{calc_prelim_stats}
\alias{calc_prelim_stats}
\title{calc_prelim_stats: Calculates initial statistics for a Seurat object
as required for downstream analysis}
\usage{
calc_prelim_stats(so, sampleID = NULL)
}
\arguments{
\item{so}{A Seurat object imported from .h5 file generated by 10X
CellRanger}

\item{sampleID}{An optional character argument to include sample label (useful
for downstream multi-sample analysis)}
}
\value{
A Seurat object with calculated statistics added as metadata
}
\description{
Calculates the percentage of mitochondrial and ribosomal
reads for each cell in the Seurat object. These statistics can be used for
filtering or for feature regression in downstream normalization.
}
