% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_metrics.R
\name{cluster_metrics}
\alias{cluster_metrics}
\title{cluster_metrics: Calculates clustering metric scores}
\usage{
cluster_metrics(
  so,
  cluster_list,
  dims = 1:20,
  reduction = "pca",
  silhouette = TRUE
)
}
\arguments{
\item{so}{A Seurat single cell RNA (scRNA) object}

\item{cluster_list}{A vector of cluster resolution names (e.g.
"SCT_snn_res.0.2")}

\item{dims}{An integer vector of the principal components to be used in
calculating the cluster metrics. May not exceed the number of principal
components available}

\item{reduction}{A string for the reduction to be used in calculating scores}

\item{silhouette}{A Boolean for whether to include silhouette scoring}
}
\value{
Returns a table of clustering metrics for each cluster resolution
selected
}
\description{
Uses the ClusterSim package to calculate metrics for cluster consistency.
These metrics do not require a priori knowledge of the truth set, which makes
them ideal for scRNA clustering.
}
\details{
Calculates three consistency scores to evaluate clustering effectiveness:
Calinski-Harabasz: Lower scores preferred.
Davies-Bouldin: Higher scores preferred.
Silhouette score: Higher scores preferred. Recommended to only run when data
has fewer than 90,000 data points
}
