% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mtx_to_h5.R
\name{convert_mtx_to_h5}
\alias{convert_mtx_to_h5}
\title{convert_mtx_to_h5: Converts the triple .mtx files from single cell
experiments to a .h5 counts file}
\usage{
convert_mtx_to_h5(sample_name, mtx_file, features_file, barcodes_file)
}
\arguments{
\item{sample_name}{A character string to be used to name the sample}

\item{mtx_file}{Path to the matrix.mtx.gz file}

\item{features_file}{Path to the features.tsv.gz file}

\item{barcodes_file}{Path to the barcodes.tsv.gz file}
}
\description{
Uses existing tools to generate .h5 files, which can be easier
to import or transfer
}
\details{
Older versions of CellRanger, in addition to some other single cell
alignment tools, such as PipSeq and DropSeq, produce a trio of files in an
output director that pointed to all the necessary features:
\itemize{
\item matrix.mtx.gz: A matrix counts file with cells in the columns and features
in the rows
\item barcodes.tsv.gz: A tab-separated file with cell/droplet barcode sequences
\item features.tsv.gz: A tab-separated file with gene names (or other features)
}
}
