% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cells.R
\name{filter_cells}
\alias{filter_cells}
\title{filter_cells: A wrapper for multiple scRNA filtering methods}
\usage{
filter_cells(
  so,
  method = "miQC",
  mads = 3,
  nCount_RNA_max = 50000,
  nCount_RNA_min = 1000,
  nFeature_RNA_max = 5000,
  nFeature_RNA_min = 200,
  percent_mt_max = 10,
  percent_mt_min = 0
)
}
\arguments{
\item{so}{Seurat object to filter}

\item{method}{Filtering method to use: "miQC" (default), "manual", or "mads"}

\item{mads}{Number of median absolute deviations to use for "mads" method (default 3)}

\item{nCount_RNA_max}{Maximum nCount_RNA threshold for "manual" method (default 500000)}

\item{nCount_RNA_min}{Minimum nCount_RNA threshold for "manual" method (default 1000)}

\item{nFeature_RNA_max}{Maximum nFeature_RNA threshold for "manual" method (default 5000)}

\item{nFeature_RNA_min}{Minimum nFeature_RNA threshold for "manual" method (default 200)}

\item{percent_mt_max}{Maximum percent_mt threshold for "manual" method (default 10)}

\item{percent_mt_min}{Minimum percent_mt threshold for "manual" method (default 0)}
}
\value{
Returns a Seurat object with an additional metadata column "keep" indicating
whether each cell is to be kept or discarded
}
\description{
Wrapper function to filter cells from a Seurat object using
method of choice. Returns a Seurat object with filtering results
}
\details{
This function filters cells from a Seurat object using one of three
methods:
\itemize{
\item "miQC": Uses the miQC package to filter cells based on a probabilistic
model based on mitochondrialcontent and number of detected features
\item "manual": Filters cells based on user-defined thresholds for nCount_RNA,
nFeature_RNA, and percent_mt
\item "mads": Filters cells based on median absolute deviations from the median for
nCount_RNA, nFeature_RNA, and percent_mt
}
}
