% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_statistic_dotplot.R
\name{make_statistic_dotplot}
\alias{make_statistic_dotplot}
\title{make_statistic_dotplot: Statistics-derived dot plot}
\usage{
make_statistic_dotplot(data, x_axis, y_axis, size, color, palette = "RdYlBu")
}
\arguments{
\item{data}{A dataframe containing the statistics of interest}

\item{x_axis}{A character string for the column name of the x-axis}

\item{y_axis}{A character string for the column name of the y-axis}

\item{size}{A character string for the column name defining size of the
bubbles/dots}

\item{color}{A character string for the column name defining the color}

\item{palette}{A character string defining a color palette from RColorBrewer
or a list of colors defining low, middle (optional), and high values}
}
\value{
A ggplot2 figure
}
\description{
A DotPlot visualization showing key statistics of a two-group
comparison
}
