% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_volcano_plot.R
\name{make_volcano_plot}
\alias{make_volcano_plot}
\title{make_volcano_plot: Optimized volcano plot for Seurat differential expression
(FindMarkers) output}
\usage{
make_volcano_plot(
  de_table,
  significant = TRUE,
  logfc = TRUE,
  pval = TRUE,
  label = 50
)
}
\arguments{
\item{de_table}{Differential expression table generated by Seurat containing
p_val, pct.1, pct.2 avg_log2FC and p_val_adj}

\item{significant}{Boolean to color genes that meet both logfc and p-value
thresholds}

\item{logfc}{Boolean to color genes meeting logfc threshold of 1.5}

\item{pval}{Boolean to color genes meeting p-value threshold of 0.05}

\item{label}{Numeric value of top n genes to label, or a character vector of
genes to label. Set label to NULL to generate unlabeled plot}
}
\value{
Returns a volcano plot as a ggplot2 object
}
\description{
Uses ggplot2 to create a generic volcano plot from the
differential expression table generated by Seurat
}
