% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_singleR_cluster.R
\name{run_singleR_cluster}
\alias{run_singleR_cluster}
\title{run_singleR_cluster: Runs SingleR on average expression of clusters}
\usage{
run_singleR_cluster(so_in, ref_file, label)
}
\arguments{
\item{so_in}{A Seurat single cell object}

\item{ref_file}{A SingleR compatible single cell reference object
(SingleCellExperiment object)}

\item{label}{A cell type label from the metadata column headers}
}
\value{
A character vector of matched cell type annotations based on
clusters
}
\description{
Wrapper function to run SingleR cell type annotation on the
current default identities of a Seurat scRNA object, e.g. clusters
}
