% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_singleR_db.R
\name{run_singleR_db}
\alias{run_singleR_db}
\title{run_singleR_db: Runs SingleR with the built-in databases from celldex}
\usage{
run_singleR_db(so_in, species)
}
\arguments{
\item{so_in}{A given Seurat single cell object}

\item{species}{Indicates human (hg19 or hg38) or mouse (mm10) references to
be used}
}
\value{
A Seurat single cell object with predicted cell type annotations
}
\description{
Uses the run_singleR function with the provided default
human or mouse references
}
