% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOT-package.R
\docType{package}
\name{SCOT-package}
\alias{SCOT}
\alias{SCOT-package}
\title{SCOT: Toolbox for single cell RNA-seq}
\description{
Designed for single-cell RNA-seq analysis with \href{https://github.com/CCBR/SINCLAIR}{SINCLAIR}.
}
\details{
See the website for more information, documentation, and examples:
\url{https://ccbr.github.io/SCOT}
}
\author{
\strong{Maintainer}: Nathan Wong \email{nathan.wong@nih.gov}

Authors:
\itemize{
  \item Nathan Wong \email{nathan.wong@nih.gov}
  \item Kelly Sovacool \email{kelly.sovacool@nih.gov} (\href{https://orcid.org/0000-0003-3283-829X}{ORCID})
  \item Vishal Koparde \email{vishal.koparde@nih.gov} (\href{https://orcid.org/0000-0001-8978-8495}{ORCID})
}

Other contributors:
\itemize{
  \item CCR Collaborative Bioinformatics Resource [copyright holder]
}

}
\keyword{internal}
