% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sc_gsea_score.R
\name{singleCellGSEAScore}
\alias{singleCellGSEAScore}
\title{singleCellGSEAScore: Calculates scores for pre-ranked GSEA}
\usage{
singleCellGSEAScore(deTable)
}
\arguments{
\item{deTable}{A differential expression table generated from the Seurat FindMarkers function. Contains the column headers \code{p_val}, \code{avg_log2FC}, \code{pct.1}, \code{pct.2}, and \code{p_val_adj}}
}
\value{
Returns a named vector of GSEA scores for genes
}
\description{
Creates a vector of scores that can be used to rank genes in order to be submitted to GSEA
}
\details{
The scoring metric uses a Seurat FindMarkers output table and incorporates the p-value and the fold change to determine magnitude, while using the percent expression values as weighting factors.  The basic formula is:

-log_10(p-value) * sign (log2FC) * max(pct.1, pct.2) + log2FC

If the p-value is recorded as zero (i.e. below machine error), the formula is adjusted

500 * sign(log2FC) * max(pct.1, pct.2) + log2FC

This formula should preserve the ranking as indicated by magnitude of fold change, while also diminishing the significance of lowly-expressed genes.
}
