% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_featurePlot.R
\name{splitFeaturePlot}
\alias{splitFeaturePlot}
\title{split_featurePlot: Splits FeaturePlot by groups}
\usage{
splitFeaturePlot(
  so,
  features,
  split_ident,
  label = FALSE,
  ncol = NA,
  nrow = NA,
  min.cutoff = NA,
  max.cutoff = NA,
  plot_image = FALSE,
  return_list = FALSE,
  slot = "scale.data",
  order = FALSE,
  reduction = NULL
)
}
\arguments{
\item{so}{A Seurat object}

\item{features}{A list of features (e.g. genes) to plot on a projection}

\item{label}{Boolean for whether to label the plot with the current Idents
value of the Seurat object}

\item{ncol}{Integer value for number of columns in plot}

\item{nrow}{Integer value for number of rows in plot}

\item{min.cutoff}{Vector of minimum cutoff values for each feature, may
specify quantile in the form of 'q##' where '##' is the quantile (eg, 1, 10)}

\item{max.cutoff}{Vector of maximum cutoff values for each feature, may
specify quantile in the form of 'q##' where '##' is the quantile (eg, 1, 10)}

\item{slot}{Data slot from which to extract values}

\item{order}{Boolean for whether to show highest feature values at the front
of the image}

\item{reduction}{Dimensionality reduction to use}

\item{splitIdent}{A metadata identity for splitting the image}

\item{plotImage}{Boolean for whether to export plots to current visualization
window}

\item{returnList}{Boolean for whether to export list of plots}
}
\value{
A list of ggplot2 plots
}
\description{
Adapts the FeaturePlot function from Seurat to split by groups and
arrange by user specification
}
