% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_VDJ_metadata.R
\name{add_VDJ_metadata}
\alias{add_VDJ_metadata}
\title{add_VDJ_metadata: Appends V(D)J alignments from CellRanger to an existing
Seurat object}
\usage{
add_VDJ_metadata(so, vdj_file)
}
\arguments{
\item{so}{A Seurat single cell RNA object}

\item{vdj_file}{A path to a V(D)J annotation directory from CellRanger}
}
\value{
A Seurat object with V(D)J metadata
}
\description{
Wrapper for djvdj to assign V(D)J alignments to cells in a
Seurat object. Can be used for TCRa/TCRb, TCRg/TCRd, and BCR
}
