% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{bubble_plot}
\alias{bubble_plot}
\title{bubble_plot: Flexible dotplot}
\usage{
bubble_plot(
  so,
  features,
  palette = "RdBu",
  assay = "SCT",
  ident = "seurat_clusters"
)
}
\description{
bubble_plot: Flexible dotplot
}
