% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_human_gene_list.R
\name{convert_human_gene_list}
\alias{convert_human_gene_list}
\title{convert_human_gene_list}
\usage{
convert_human_gene_list(genes)
}
\arguments{
\item{genes}{Character vector of human genes}
}
\value{
Character vector of mapped mouse genes
}
\description{
Uses the ENSEMBL database to convert human gene names to mouse gene names
}
