% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_celldex_ref.R
\name{fetch_celldex_ref}
\alias{fetch_celldex_ref}
\title{fetch_celldex_ref: Helper function to import annotation references from
celldex}
\usage{
fetch_celldex_ref(ref_name)
}
\value{
A celldex single cell object in SCE data structure
}
\description{
Retrieves one of the 7 available references in celldex. Should
be cache-agnostic by declaring the cache location as the local working
directory
}
