% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_doublets.R
\name{filter_doublets}
\alias{filter_doublets}
\title{filter_doublets: Identifies and removes doublets or multiplets in scRNA
samples}
\usage{
filter_doublets(so_in, doublet_finder_method = "DoubletFinder")
}
\arguments{
\item{so_in}{A provided Seurat scRNA object}

\item{doublet_finder_method}{Character string indicating the use of
"DoubletFinder" (default), "scDblFinder", "consensus" removal of doublets,
or the "union" of total doublets identified in both algorithms}
}
\description{
Wrapper for DoubletFinder and scDblFinder for identifying doublets
}
\details{
Runs the latest version of DoubletFinder and/or scDblFinder and returns
a filtered scRNA object with doublets removed
}
