% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bubble_plot.R
\name{make_bubble_plot}
\alias{make_bubble_plot}
\title{make_bubble_plot: Flexible dotplot}
\usage{
make_bubble_plot(
  so,
  features,
  palette = "RdBu",
  assay = "SCT",
  ident = "seurat_clusters"
)
}
\arguments{
\item{so}{A Seurat single cell RNA object}

\item{features}{A character vector of genes to plot}

\item{palette}{A color palette from ggplot2}

\item{assay}{The counts assay to use for determining expression}

\item{ident}{The categorical identity to classify groups of cells}
}
\value{
A ggplot2 figure
}
\description{
Creates a bubble plot that shows average expression percent
expression for genes in single cell categories
}
