% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sample.R
\name{preprocess_sample}
\alias{preprocess_sample}
\title{preprocess_sample: Performs an initial preprocess on a Seurat object}
\usage{
preprocess_sample(so_in, species, npcs_in)
}
\arguments{
\item{so_in}{A Seurat single cell object with RNA counts}

\item{species}{Species of the sample (hg38, hg19, or mm10)}

\item{npcs_in}{Number of principal components to use for dimensionality
reduction}
}
\value{
A Seurat single cell object with normalized read counts
}
\description{
: Conducts preliminary normalization/scaling, cell cycle scoring,
and clustering on a Seurat single cell object
}
\details{
The function performs the following processes on the data:assay
\itemize{
\item Cell cycle scoring
\item Normalization and scaling of RNA counts with log normalization
\item Normalization and scaling of RNA counts with SCTransform
\item Principal components analysis, neighbor identification, UMAP dimensionality
reduction and preliminary clustering (through \code{seurat_clustering} function)
}
}
