% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_singleR.R
\name{run_singleR}
\alias{run_singleR}
\title{run_singleR: Runs SingleR directly on the Seurat object}
\usage{
run_singleR(so, ref_file, label)
}
\arguments{
\item{so}{A Seurat single cell object}

\item{ref_file}{A reference file in a SingleCellExperiment format, such as those obtained from SingleR/CellDex
package}

\item{label}{The label identity to be used. Must be a column header in the metadata of \code{ref_file}}
}
\value{
A vector of pruned cell type labels
}
\description{
Wrapper script for the SingleR function
}
