% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitFeaturePlot.R
\name{splitFeaturePlot}
\alias{splitFeaturePlot}
\title{splitFeaturePlot: Splits FeaturePlot by groups}
\usage{
splitFeaturePlot(
  so,
  features,
  splitIdent,
  label = F,
  ncol = NA,
  nrow = NA,
  min.cutoff = NA,
  max.cutoff = NA,
  plotImage = F,
  returnList = F,
  slot = "scale.data",
  order = F,
  reduction = NULL
)
}
\arguments{
\item{so}{A Seurat object}

\item{features}{A list of features (e.g. genes) to plot on a projection}

\item{splitIdent}{A metadata identity for splitting the image}

\item{label}{Boolean for whether to label the plot with the current Idents value of the Seurat object}

\item{ncol}{Integer value for number of columns in plot}

\item{nrow}{Integer value for number of rows in plot}

\item{min.cutoff}{Vector of minimum cutoff values for each feature, may specify quantile in the form of 'q##' where '##' is the quantile (eg, 1, 10)}

\item{max.cutoff}{Vector of maximum cutoff values for each feature, may specify quantile in the form of 'q##' where '##' is the quantile (eg, 1, 10)}

\item{plotImage}{Boolean for whether to export plots to current visualization window}

\item{returnList}{Boolean for whether to export list of plots}

\item{slot}{Data slot from which to extract values}

\item{order}{Boolean for whether to show highest feature values at the front of the image}

\item{reduction}{Dimensionality reduction to use}
}
\value{
A list of ggplot2 plots
}
\description{
Adapts the FeaturePlot function from Seurat to split by groups and arrange by user specification
}
