% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{volcano_plot: Optimized volcano plot for Seurat differential expression (FindMarkers) output}
\usage{
volcano_plot(de_table, significant = TRUE, logfc = TRUE, pval = TRUE)
}
\arguments{
\item{de_table}{Differential expression table generated by Seurat containing p_val, pct.1, pct.2 avg_log2FC and p_val_adj}

\item{significant}{Boolean to color genes that meet both above thresholds}

\item{logfc}{Boolean to color genes meeting logfc threshold of 1.5}

\item{pval}{Boolean to color genes meeting p-value threshold of 0.05}
}
\value{
Returns a volcano plot as a ggplot2 object
}
\description{
Uses ggplot2 to create a generic volcano plot from the differential expression table generated by Seurat
}
